split emit target selection validation tests
