+++ docs/PHASE_PLAN.md
@@ -464,6 +464,9 @@ checked-in docs, tests, and commits only.
 - Build-graph JSON validation tests now keep unsupported target kind and gated
   package/link field diagnostics in a focused child module, leaving target
   field and dependency graph diagnostics in the parent module.
+- Emit-command build graph validation tests now keep target-selection
+  diagnostics in a focused child module, leaving unrelated gated test source
+  skip coverage in the parent module.
 - Generic specialization generated-C assertions now keep multi-file import and
   dependency cases in a focused integration module, separating them from
   single-file generic specialization emission assertions.
